src/requirements.txt
Babel==2.5.3
Django==1.9
django-debug-toolbar==1.9.1
MySQL-python==1.2.5
olefile==0.44
Pillow==4.3.0
pytz==2017.3
sqlparse==0.2.4
